backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
sqlalchemy==2.0.23
geoalchemy2==0.14.3
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
opencv-python-headless==4.9.0.80
pillow==10.1.0
google-cloud-storage==2.14.0
cloudinary==1.36.0
python-dotenv==1.0.0
